beanstalk/ebs/requirements.txt
#
####### python requirements #######
#
###### Requirements without Version Specifiers ######
boto3
httplib2
awsebcli
botocore
awscli